// lib/smart-onboarding/config/database.d.ts
export declare const smartOnboardingDb: {
    query: (text: string, params?: any[]) => Promise<import("pg").QueryResult<any>>;
    getPool: typeof getPool;
    tables: {
        readonly USER_PROFILES: "smart_onboarding_user_profiles";
        readonly ONBOARDING_JOURNEYS: "smart_onboarding_journeys";
        readonly ONBOARDING_STEPS: "smart_onboarding_steps";
        readonly LEARNING_PATHS: "smart_onboarding_learning_paths";
        readonly BEHAVIOR_EVENTS: "smart_onboarding_behavior_events";
        readonly INTERACTION_DATA: "smart_onboarding_interaction_data";
        readonly ENGAGEMENT_METRICS: "smart_onboarding_engagement_metrics";
        readonly STRUGGLE_INDICATORS: "smart_onboarding_struggle_indicators";
        readonly USER_PERSONAS: "smart_onboarding_user_personas";
        readonly PREDICTION_MODELS: "smart_onboarding_prediction_models";
        readonly MODEL_TRAINING_DATA: "smart_onboarding_model_training_data";
        readonly SUCCESS_PREDICTIONS: "smart_onboarding_success_predictions";
        readonly INTERVENTIONS: "smart_onboarding_interventions";
        readonly INTERVENTION_OUTCOMES: "smart_onboarding_intervention_outcomes";
        readonly HELP_CONTENT: "smart_onboarding_help_content";
        readonly ESCALATION_TICKETS: "smart_onboarding_escalation_tickets";
        readonly CONTENT_VARIATIONS: "smart_onboarding_content_variations";
        readonly PERSONALIZED_CONTENT: "smart_onboarding_personalized_content";
        readonly CONTENT_RECOMMENDATIONS: "smart_onboarding_content_recommendations";
        readonly ADAPTATION_HISTORY: "smart_onboarding_adaptation_history";
        readonly BEHAVIORAL_INSIGHTS: "smart_onboarding_behavioral_insights";
        readonly PATH_PERFORMANCE: "smart_onboarding_path_performance";
        readonly EXPERIMENT_RESULTS: "smart_onboarding_experiment_results";
        readonly SYSTEM_METRICS: "smart_onboarding_system_metrics";
    };
    cacheKeys: {
        readonly USER_PROFILE: (userId: string) => string;
        readonly USER_JOURNEY: (userId: string) => string;
        readonly USER_PERSONA: (userId: string) => string;
        readonly ENGAGEMENT_SCORE: (userId: string) => string;
        readonly LEARNING_PATH: (pathId: string) => string;
        readonly STEP_CONTENT: (stepId: string) => string;
        readonly CONTENT_RECOMMENDATIONS: (userId: string, stepId: string) => string;
        readonly MODEL_PREDICTIONS: (userId: string, modelType: string) => string;
        readonly SUCCESS_PREDICTION: (userId: string) => string;
        readonly SYSTEM_METRICS: "smart_onboarding:system_metrics";
        readonly ACTIVE_EXPERIMENTS: "smart_onboarding:active_experiments";
        readonly MODEL_PERFORMANCE: (modelId: string) => string;
    };
    cacheTtl: {
        readonly ENGAGEMENT_SCORE: 60;
        readonly BEHAVIOR_EVENTS: 300;
        readonly REAL_TIME_METRICS: 30;
        readonly USER_JOURNEY: 1800;
        readonly CONTENT_RECOMMENDATIONS: 900;
        readonly INTERVENTION_PLANS: 600;
        readonly USER_PROFILE: 3600;
        readonly USER_PERSONA: 7200;
        readonly LEARNING_PATH: 14400;
        readonly STEP_CONTENT: 21600;
        readonly MODEL_PREDICTIONS: 1800;
        readonly SUCCESS_PREDICTION: 3600;
        readonly MODEL_PERFORMANCE: 86400;
        readonly SYSTEM_METRICS: 300;
        readonly ACTIVE_EXPERIMENTS: 3600;
    };
};